pocl: disable bottles
The latest pocl (0.13) fails to build with the current llvm formula
(3.8.1) as well as with the prior (3.8.0):

```
error "isfinite is not a macro"
```

pocl 0.12 is not compatible with llvm 3.8*, but it does build with our
prior major llvm version, which was llvm (3.6.2), but this version is no
longer available in core and the version in homebrew/dupes differs.
Also, despite the successful build, there is still a test failure in
`make check`:

```
 16: Kernel functions length, distance, and normalize FAILED
(testsuite.at:174)
```

pocl 0.11 also builds with llvm 3.6.2, and `make check` succeeds, but
this may simply be because the test that failed for 0.12 appears to be
new.